# OPENAPI_SPEC.yaml
openapi: 3.1.0
info:
  title: Project Management Orchestrator - Revised
  version: 2.0.0
  description: |
    Unified API with proper organization isolation and role-based access control.
    
    **Authentication**: All endpoints require Bearer token except /register and /login
    **Authorization**: Enforced at endpoint level based on user role
    **Data Isolation**: Users can only access data from their organization
servers:
  - url: http://localhost:9000
    description: Development server
  - url: https://api.example.com
    description: Production server

# ============================================
# AUTHENTICATION & ORGANIZATION MANAGEMENT
# ============================================
paths:
...
  /api/v1/activities:
    post:
      tags: [Activities]
      summary: Log activity
      description: Log work activity with optional task link
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, duration_seconds]
              properties:
                description:
                  type: string
                duration_seconds:
                  type: integer
                  minimum: 0
                task_id:
                  type: string
                  format: uuid
                  nullable: true
                logged_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Activity logged
        '400':
          description: Task not in user's organization (if task_id provided)

    get:
      tags: [Activities]
      summary: Get user's activities
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of activities
  /api/v1/activities/team:
    get:
      tags: [Activities]
      summary: Get team activities
      description: ⚠️ Only admin or manager can view team activities
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Team activity summary
        '403':
          description: Insufficient permissions

# ============================================
# TEAMS (WorkPulse Service)
# ============================================
  /api/v1/teams:
    post:
      tags: [Teams]
      summary: Create team
      description: ⚠️ Only admin or manager can create teams
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                leader_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: Must be org member
      responses:
        '201':
          description: Team created
        '403':
          description: Only admin or manager can create teams

    get:
      tags: [Teams]
      summary: List teams in organization
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of teams
  /api/v1/teams/{team_id}/members:
    post:
      tags: [Teams]
      summary: Add member to team
      description: ⚠️ Only admin, manager, or team leader can add members
      security:
        - bearerAuth: []
      parameters:
        - name: team_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                  format: uuid
                  description: Must be org member
                role:
                  type: string
                  enum: [member, lead]
                  default: member
      responses:
        '201':
          description: Member added to team
        '403':
          description: Insufficient permissions
        '400':
          description: User not in organization or already in team

# ============================================
# PERFORMANCE (EPR Service)
# ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...